Fixed recurring error in MainActivity.java. It could not find the ArticlePreview class.
