+++ NewsAppV3/app/src/main/java/com/example/newsappv3/SearchArticlePreview.java
@@ -35,7 +35,7 @@ public class SearchArticlePreview extends NewsDriver {
         listNews.setEmptyView(loader);
 
         // Download news if network is available
-        if (SharedFunctions.isNetworkAvailable(getApplicationContext())) {
+        if (SharedResources.isNetworkAvailable(getApplicationContext())) {
             DownloadNews newsTask = new DownloadNews();
             newsTask.execute();
         } else {
@@ -74,7 +74,7 @@ public class SearchArticlePreview extends NewsDriver {
 
         // Perform a Get on the Given URL
         protected String doInBackground(String... args) {
-            String xml = SharedFunctions.excuteGet("https://newsapi.org/v2/everything?" + "q=" + keyWord + "&apiKey=" + API_KEY);
+            String xml = SharedResources.excuteGet("https://newsapi.org/v2/everything?" + "q=" + keyWord + "&apiKey=" + API_KEY);
             return xml;
         }
 
@@ -92,12 +92,13 @@ public class SearchArticlePreview extends NewsDriver {
                     for (int i = 0; i < jsonArray.length(); i++) {
                         JSONObject jsonObject = jsonArray.getJSONObject(i);
                         HashMap<String, String> map = new HashMap<>();
-                        map.put(KEY_AUTHOR, jsonObject.optString(KEY_AUTHOR));
+                        JSONObject source = jsonObject.getJSONObject(KEY_SOURCE);
+                        map.put(KEY_NAME, source.getString(KEY_NAME));
                         map.put(KEY_TITLE, jsonObject.optString(KEY_TITLE));
                         map.put(KEY_DESCRIPTION, jsonObject.optString(KEY_DESCRIPTION));
                         map.put(KEY_URL, jsonObject.optString(KEY_URL));
                         map.put(KEY_URLTOIMAGE, jsonObject.optString(KEY_URLTOIMAGE));
-                        map.put(KEY_PUBLISHEDAT, jsonObject.optString(KEY_PUBLISHEDAT));
+                        map.put(KEY_PUBLISHEDAT, SharedResources.DateFormat(jsonObject.optString(KEY_PUBLISHEDAT)));
                         dataList.add(map);
                     }
                 } catch (JSONException e) {
@@ -105,7 +106,7 @@ public class SearchArticlePreview extends NewsDriver {
                 }
 
                 // Populate the Article preview view
-                ArticlePreview adapter = new ArticlePreview(SearchArticlePreview.this, dataList);
+                SharedResources.ArticlePreview adapter = new SharedResources.ArticlePreview(SearchArticlePreview.this, dataList);
                 listNews.setAdapter(adapter);
 
                 // Listener for Article clicks
