--- NewsAppV3/app/src/main/java/com/example/newsappv3/SharedFunctions.java
@@ -1,90 +1,0 @@
-package com.example.newsappv3;
-
-// Contains functions that are common in both the Main page and Search Page
-// Methods to check for network connection and perform a GET on a given url
-
-import android.content.Context;
-import android.net.ConnectivityManager;
-import java.io.BufferedReader;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.util.Locale;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-
-
-public class SharedFunctions {
-
-    // Check in network is available
-    public static boolean isNetworkAvailable(Context context) {
-        return ((ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE)).getActiveNetworkInfo() != null;
-    }
-
-    // Call news endpoint and save the response
-    public static String excuteGet(String targetURL) {
-        URL url;
-        HttpURLConnection connection = null;
-        try {
-            // Make the http connection
-            url = new URL(targetURL);
-            connection = (HttpURLConnection) url.openConnection();
-            connection.setRequestProperty("content-type", "application/json;  charset=utf-8");
-            connection.setRequestProperty("Content-Language", "en-US");
-            connection.setUseCaches(false);
-            connection.setDoInput(true);
-            connection.setDoOutput(false);
-
-            InputStream input;
-
-            // Check that connection was made properly
-            int status = connection.getResponseCode();
-            if (status != HttpURLConnection.HTTP_OK)
-                input = connection.getErrorStream();
-            else
-                // if connection was successful then save data
-                input = connection.getInputStream();
-
-            // Save each article from the input stream into a StringBuffer
-            BufferedReader reader = new BufferedReader(new InputStreamReader(input));
-            String line;
-            StringBuffer response = new StringBuffer();
-            while ((line = reader.readLine()) != null) {
-                response.append(line);
-                response.append('\r');
-            }
-            reader.close();
-            return response.toString();
-        } catch (Exception e) {
-            return null;
-        } finally {
-            if (connection != null) {
-                connection.disconnect();
-            }
-        }
-    }
-
-    // Reformat a given date
-    public static String DateFormat(String oldstringDate){
-        String newDate;
-        SimpleDateFormat dateFormat = new SimpleDateFormat("E, d MMM yyyy", new Locale(getCountry()));
-        try {
-            Date date = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'").parse(oldstringDate);
-            newDate = dateFormat.format(date);
-        } catch (ParseException e) {
-            e.printStackTrace();
-            newDate = oldstringDate;
-        }
-
-        return newDate;
-    }
-
-    public static String getCountry(){
-        Locale locale = Locale.getDefault();
-        String country = String.valueOf(locale.getCountry());
-        return country.toLowerCase();
-    }
-}
-
