+++ NewsAppV3/app/src/main/res/layout/article_info.xml
@@ -75,12 +75,12 @@
                 android:orientation="horizontal">
 
                 <TextView
-                    android:id="@+id/author"
+                    android:id="@+id/name"
                     android:textSize="15sp"
                     android:textColor="#42A5F5"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="News Author"/>
+                    android:text="Name of source"/>
 
             </LinearLayout>
